+++ instrumentation/runtime/httpurlconnection-auto/build.gradle.kts
@@ -25,6 +25,8 @@ android {
 dependencies {
     implementation(platform(Dependencies.Otel.instrumentationBomAlpha))
 
+    implementation(project(":common:utils"))
+
     implementation(Dependencies.Otel.instrumentationApi)
     implementation(Dependencies.Otel.instrumentationApiIncubator)
