+++ instrumentation/runtime/httpurlconnection-auto/src/main/java/com/splunk/rum/instrumentation/httpurlconnection/auto/internal/PeerServiceAttributesExtractor.kt
@@ -17,7 +17,7 @@
 
 package com.splunk.rum.instrumentation.httpurlconnection.auto.internal
 
-import com.splunk.rum.instrumentation.httpurlconnection.auto.HttpUrlInstrumentation
+import com.splunk.rum.agent.common.utils.PeerServiceMappingResolver
 import io.opentelemetry.api.common.AttributeKey
 import io.opentelemetry.api.common.AttributesBuilder
 import io.opentelemetry.context.Context
@@ -26,13 +26,7 @@ import io.opentelemetry.instrumentation.api.semconv.http.HttpClientAttributesGet
 import java.net.URI
 import java.net.URLConnection
 
-/**
- * Preserves the programmatic `peer.service` mapping supported by the previous instrumentation API.
- *
- * The replacement upstream service-peer extractor only reads declarative configuration and cannot
- * consume [HttpUrlInstrumentation.setPeerServiceMapping] values. Keeping this extractor local
- * avoids changing the SDK's telemetry schema during the dependency upgrade.
- */
+/** Extracts the configured `peer.service` mapping for HttpURLConnection requests. */
 internal class PeerServiceAttributesExtractor(
     private val attributesGetter: HttpClientAttributesGetter<URLConnection, Int>,
     peerServiceMapping: Map<String, String>
@@ -48,78 +42,17 @@ internal class PeerServiceAttributesExtractor(
         response: Int?,
         error: Throwable?
     ) {
-        if (resolver.isEmpty()) {
-            return
-        }
-
+        val path = runCatching { URI(attributesGetter.getUrlFull(request)).path }.getOrNull()
         val serviceName = resolver.resolve(
             attributesGetter.getServerAddress(request),
             attributesGetter.getServerPort(request),
-            attributesGetter.getUrlFull(request).toPathOrNull()
+            path
         ) ?: return
 
         attributes.put(PEER_SERVICE, serviceName)
     }
 
-    private class PeerServiceMappingResolver(peerServiceMapping: Map<String, String>) {
-        private val mappingsByHost: Map<String, List<Mapping>> =
-            peerServiceMapping.entries
-                .mapNotNull { (peer, serviceName) -> Mapping.parse(peer, serviceName) }
-                .groupBy(Mapping::host)
-
-        fun isEmpty(): Boolean = mappingsByHost.isEmpty()
-
-        fun resolve(host: String?, port: Int?, path: String?): String? {
-            if (host == null) {
-                return null
-            }
-
-            return mappingsByHost[host]
-                ?.asSequence()
-                ?.filter { it.matches(port, path) }
-                ?.maxWithOrNull(MAPPING_SPECIFICITY)
-                ?.serviceName
-        }
-    }
-
-    private data class Mapping(val host: String, val port: Int?, val path: String?, val serviceName: String) {
-        fun matches(requestPort: Int?, requestPath: String?): Boolean {
-            if (port != null && port != requestPort) {
-                return false
-            }
-
-            if (!path.isNullOrEmpty()) {
-                if (requestPath == null || !requestPath.startsWith(path)) {
-                    return false
-                }
-                if (requestPort != null && requestPort != port) {
-                    return false
-                }
-            }
-
-            return true
-        }
-
-        companion object {
-            fun parse(peer: String, serviceName: String): Mapping? = runCatching {
-                val uri = URI("https://$peer")
-                Mapping(
-                    host = uri.host ?: return null,
-                    port = uri.port.takeIf { it >= 0 },
-                    path = uri.path,
-                    serviceName = serviceName
-                )
-            }.getOrNull()
-        }
-    }
-
     private companion object {
         private val PEER_SERVICE = AttributeKey.stringKey("peer.service")
-
-        private val MAPPING_SPECIFICITY =
-            compareBy<Mapping, Int?>(nullsFirst(naturalOrder())) { it.port }
-                .thenBy(nullsFirst(naturalOrder())) { it.path }
-
-        private fun String?.toPathOrNull(): String? = this?.let { url -> runCatching { URI(url).path }.getOrNull() }
     }
 }
